pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoapply"
version = "1.0.0"
description = "AI-assisted resume tailoring and application orchestrator (Python rewrite)"
requires-python = ">=3.11"
dependencies = [
  # Core framework
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "questionary>=2.0",
  
  # Database & ORM
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.12",
  
  # Document parsing
  "PyPDF2>=3.0",
  "pdfplumber>=0.10",
  "python-docx>=1.0",
  
  # Web scraping
  "beautifulsoup4>=4.12",
  "requests>=2.31",
  "lxml>=5.0",
  
  # Document generation
  "docx2pdf>=0.1",
  
  # Caching
  "redis>=5.0",
  
  # AI providers
  "anthropic>=0.25",
  "openai>=1.0",
  "google-generativeai>=0.5",
  
  # Async support
  "aiofiles>=23.0",
  "httpx>=0.27",
  
  # Security & encryption
  "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "black>=24.0",
  "ruff>=0.3",
  "mypy>=1.9",
]

[project.scripts]
autoapply = "autoapply.cli.app:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
